fix: cast t.speed from float32 to float64 in TextToSpeech function

+++ internal/ai/tts.go
@@ -74,7 +74,7 @@ func (t *TTSClient) TextToSpeech(text string, outputPath string) error {
 	default:
 		req.Voice = openai.VoiceAlloy
 	}
-	req.Speed = t.speed
+	req.Speed = float64(t.speed)
 
 	// Make the request
 	response, err := t.client.CreateSpeech(ctx, req)
